add integration tests and address comments
