+++ src/prime_rl/inference/vllm/worker/kernels/__init__.py
@@ -1,0 +1,3 @@
+from prime_rl.inference.vllm.worker.kernels.fp8 import quantize_weight_to_block_fp8
+
+__all__ = ["quantize_weight_to_block_fp8"]
